+++ module/account/src/Infrastructure/Persistence/Repository/EventStoreRoleRepository.php
@@ -17,7 +17,7 @@ use Ergonode\EventSourcing\Infrastructure\Manager\EventStoreManager;
 use Ergonode\SharedKernel\Domain\Aggregate\RoleId;
 use Webmozart\Assert\Assert;
 
-class DbalRoleRepository implements RoleRepositoryInterface
+class EventStoreRoleRepository implements RoleRepositoryInterface
 {
     private EventStoreManager $manager;
