+++ module/account/src/Infrastructure/Persistence/Repository/EventStoreUserRepository.php
@@ -15,7 +15,7 @@ use Ergonode\EventSourcing\Infrastructure\Manager\EventStoreManager;
 use Ergonode\SharedKernel\Domain\Aggregate\UserId;
 use Webmozart\Assert\Assert;
 
-class DbalUserRepository implements UserRepositoryInterface
+class EventStoreUserRepository implements UserRepositoryInterface
 {
     private EventStoreManager $manager;
