+++ module/account/src/Resources/config/services.yml
@@ -46,8 +46,8 @@ services:
         arguments:
             $hosts: '%ergonode_account.hosts%'
 
-    Ergonode\Account\Domain\Repository\UserRepositoryInterface: '@Ergonode\Account\Infrastructure\Persistence\Repository\DbalUserRepository'
-    Ergonode\Account\Domain\Repository\RoleRepositoryInterface: '@Ergonode\Account\Infrastructure\Persistence\Repository\DbalRoleRepository'
+    Ergonode\Account\Domain\Repository\UserRepositoryInterface: '@Ergonode\Account\Infrastructure\Persistence\Repository\EventStoreUserRepository'
+    Ergonode\Account\Domain\Repository\RoleRepositoryInterface: '@Ergonode\Account\Infrastructure\Persistence\Repository\EventStoreRoleRepository'
     Ergonode\Account\Domain\Query\AccountQueryInterface: '@Ergonode\Account\Infrastructure\Persistence\Query\DbalAccountQuery'
     Ergonode\Account\Domain\Query\ProfileQueryInterface: '@Ergonode\Account\Infrastructure\Persistence\Query\DbalProfileQuery'
     Ergonode\Account\Domain\Query\PrivilegeQueryInterface: '@Ergonode\Account\Infrastructure\Persistence\Query\DbalPrivilegeQuery'
