+++ module/attribute/src/Infrastructure/Persistence/Repository/EventStoreAttributeRepository.php
@@ -16,7 +16,7 @@ use Ergonode\EventSourcing\Infrastructure\Manager\EventStoreManager;
 use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use Webmozart\Assert\Assert;
 
-class DbalAttributeRepository implements AttributeRepositoryInterface
+class EventStoreAttributeRepository implements AttributeRepositoryInterface
 {
     private EventStoreManager $manager;
