+++ module/attribute/src/Infrastructure/Persistence/Repository/EventStoreOptionRepository.php
@@ -17,7 +17,7 @@ use Ergonode\EventSourcing\Infrastructure\Manager\EventStoreManager;
 use Ergonode\SharedKernel\Domain\AggregateId;
 use Webmozart\Assert\Assert;
 
-class DbalOptionRepository implements OptionRepositoryInterface
+class EventStoreOptionRepository implements OptionRepositoryInterface
 {
     private EventStoreManager $manager;
