+++ module/category/src/Infrastructure/Persistence/Repository/EventStoreTreeRepository.php
@@ -17,7 +17,7 @@ use Ergonode\EventSourcing\Infrastructure\Manager\EventStoreManager;
 use Ergonode\SharedKernel\Domain\Aggregate\CategoryTreeId;
 use Webmozart\Assert\Assert;
 
-class DbalTreeRepository implements TreeRepositoryInterface
+class EventStoreTreeRepository implements TreeRepositoryInterface
 {
     private EventStoreManager $manager;
