+++ module/condition/src/Infrastructure/Persistence/Repository/EventStoreConditionSetRepository.php
@@ -16,7 +16,7 @@ use Ergonode\EventSourcing\Infrastructure\Manager\EventStoreManager;
 use Ergonode\SharedKernel\Domain\Aggregate\ConditionSetId;
 use Webmozart\Assert\Assert;
 
-class DbalConditionSetRepository implements ConditionSetRepositoryInterface
+class EventStoreConditionSetRepository implements ConditionSetRepositoryInterface
 {
     private EventStoreManager $manager;
