+++ module/designer/src/Infrastructure/Persistence/Repository/EventStoreTemplateGroupRepository.php
@@ -16,7 +16,7 @@ use Ergonode\EventSourcing\Infrastructure\Manager\EventStoreManager;
 use Ergonode\SharedKernel\Domain\Aggregate\TemplateGroupId;
 use Webmozart\Assert\Assert;
 
-class DbalTemplateGroupRepository implements TemplateGroupRepositoryInterface
+class EventStoreTemplateGroupRepository implements TemplateGroupRepositoryInterface
 {
     private EventStoreManager $manager;
