+++ module/designer/src/Infrastructure/Persistence/Repository/EventStoreTemplateRepository.php
@@ -17,7 +17,7 @@ use Ergonode\EventSourcing\Infrastructure\Manager\EventStoreManager;
 use Ergonode\SharedKernel\Domain\Aggregate\TemplateId;
 use Webmozart\Assert\Assert;
 
-class DbalTemplateRepository implements TemplateRepositoryInterface
+class EventStoreTemplateRepository implements TemplateRepositoryInterface
 {
     private EventStoreManager $manager;
