+++ module/importer/src/Resources/config/services.yml
@@ -44,7 +44,7 @@ services:
         arguments:
             $targetDirectory: '%import_directory%'
 
-    Ergonode\Importer\Domain\Repository\TransformerRepositoryInterface: '@Ergonode\Importer\Infrastructure\Persistence\Repository\DbalTransformerRepository'
+    Ergonode\Importer\Domain\Repository\TransformerRepositoryInterface: '@Ergonode\Importer\Infrastructure\Persistence\Repository\EventStoreTransformerRepository'
 
     Ergonode\Importer\Infrastructure\Action\Process\Product\ImportProductAttributeBuilder:
         arguments:
