+++ module/multimedia/src/Infrastructure/Persistence/Repository/EventStoreMultimediaRepository.php
@@ -17,7 +17,7 @@ use Ergonode\Multimedia\Domain\Repository\MultimediaRepositoryInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\MultimediaId;
 use Webmozart\Assert\Assert;
 
-class DbalMultimediaRepository implements MultimediaRepositoryInterface
+class EventStoreMultimediaRepository implements MultimediaRepositoryInterface
 {
     private EventStoreManager $manager;
