+++ module/segment/src/Infrastructure/Persistence/Repository/EventStoreSegmentRepository.php
@@ -16,7 +16,7 @@ use Ergonode\Segment\Domain\Repository\SegmentRepositoryInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\SegmentId;
 use Webmozart\Assert\Assert;
 
-class DbalSegmentRepository implements SegmentRepositoryInterface
+class EventStoreSegmentRepository implements SegmentRepositoryInterface
 {
     private EventStoreManager $manager;
